pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oferta-formativa-etl"
version = "0.1.0"
description = "ETL pipeline for the Oferta Formativa Chile 2025 dataset"
readme = "README.md"
authors = [
  {name = "Bruno San Martin Navarro"}
]
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "pyyaml>=6.0",
  "pyarrow>=17.0",
  "plotly>=5.24",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
oferta-formativa-etl = "oferta_formativa_etl.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
